Removed the modal context and made use of redux modal states
