+++ frontend/src/app/layout.tsx
@@ -1,6 +1,5 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import ModalContextLayout from "@/components/common/ModalContextLayout";
 import { Toaster } from "react-hot-toast";
 import ProviderLayout from "@/components/common/ProviderLayout";
 export const metadata: Metadata = {
@@ -23,11 +22,9 @@ export default function RootLayout({
         <Toaster
         />
         <ProviderLayout>
-          <ModalContextLayout>
-            {
-              children
-            }
-          </ModalContextLayout>
+          {
+            children
+          }
         </ProviderLayout>
