+++ frontend/src/services/productApiSlice.ts
@@ -1,4 +1,5 @@
-import { PRODUCTS_URL } from "@/constant";
+
+import { PRODUCTS_URL } from "@/constants";
 import { apiSlice } from "./apiSlice";
 
 export const productApiSlice = apiSlice.injectEndpoints({
